+++ src/MachinaDynamo/Programs/_graveyard.cs
@@ -1,0 +1,61 @@
+﻿using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+using Autodesk.DesignScript.Runtime;
+using Autodesk.DesignScript.Interfaces;
+using Autodesk.DesignScript.Geometry;
+
+using Machina;
+using MAction = Machina.Action;
+using MPoint = Machina.Point;
+using MVector = Machina.Vector;
+using MOrientation = Machina.Orientation;
+using MTool = Machina.Tool;
+using System.IO;
+using System.Collections.Generic;
+using System;
+
+using WebSocketSharp;
+using WebSocketSharp.Net;
+using WebSocketSharp.Net.WebSockets;
+using WebSocketSharp.Server;
+
+namespace MachinaDynamo
+{
+
+    //   ██████╗ ██████╗  █████╗ ██╗   ██╗███████╗██╗   ██╗ █████╗ ██████╗ ██████╗ 
+    //  ██╔════╝ ██╔══██╗██╔══██╗██║   ██║██╔════╝╚██╗ ██╔╝██╔══██╗██╔══██╗██╔══██╗
+    //  ██║  ███╗██████╔╝███████║██║   ██║█████╗   ╚████╔╝ ███████║██████╔╝██║  ██║
+    //  ██║   ██║██╔══██╗██╔══██║╚██╗ ██╔╝██╔══╝    ╚██╔╝  ██╔══██║██╔══██╗██║  ██║
+    //  ╚██████╔╝██║  ██║██║  ██║ ╚████╔╝ ███████╗   ██║   ██║  ██║██║  ██║██████╔╝
+    //   ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═╝  ╚═══╝  ╚══════╝   ╚═╝   ╚═╝  ╚═╝╚═╝  ╚═╝╚═════╝ 
+    //                                                                             
+    /// <summary>
+    /// Program creation and visualization options.
+    /// </summary>
+    public partial class Programs
+    {
+        /// <summary>
+        /// Writes a list of strings to a file. Make sure Dynamo has the adequate Admin rights to write files to your system. 
+        /// </summary>
+        /// <param name="code">A List of Strings</param>
+        /// <param name="filepath">The path where the file will be saved</param>
+        /// <returns name="resultMsg">Success?</returns>
+        [IsVisibleInDynamoLibrary(false)]
+        internal static string WriteToFile(List<string> code, string filepath)
+        {
+            string result;
+            try
+            {
+                System.IO.File.WriteAllLines(filepath, code, System.Text.Encoding.ASCII);
+                result = "Successfuly saved to " + filepath;
+            }
+            catch (Exception ex)
+            {
+                result = "Could not save to file " + filepath + ", ERROR: " + ex;
+            }
+            return result;
+        }
+    }
+}